moved keybindings to gui interfaces
